build: resolve cycle dependences
