+++ src/components/record/Airdrop.tsx
@@ -6,9 +6,12 @@ import { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { map } from 'rxjs';
 import Web3 from 'web3';
-import { crabConfig, DATE_TIME_FORMATE, FORM_CONTROL, NETWORK_LIGHT_THEME, SubscanApiPath } from '../../config';
+import { SubscanApiPath } from '../../config/api';
+import { DATE_TIME_FORMATE, FORM_CONTROL } from '../../config/constant';
+import { crabConfig } from '../../config/network';
+import { NETWORK_LIGHT_THEME } from '../../config/theme';
 import { airportsDepartureFilter, useApi, useNetworks } from '../../hooks';
-import { ClaimsRes, ChainConfig, Network, SubscanResponse } from '../../model';
+import { ChainConfig, ClaimsRes, Network, SubscanResponse } from '../../model';
 import { apiUrl, fromWei, getAirdropData, getInitialSetting, isSameNetConfig, rxGet } from '../../utils';
 import { Destination } from '../form-control/Destination';
 import { ConnectionIndicator } from '../widget/ConnectionIndicator';
