+++ src/config/constant.ts
@@ -18,3 +18,12 @@ export enum FORM_CONTROL {
   sender = 'sender',
   direction = 'direction',
 }
+
+export enum Path {
+  root = '/',
+  history = '/history',
+  airdrop = '/airdrop',
+  airdropHistory = '/airdropHistory',
+  register = '/register',
+  configure = '/configure',
+}
