+++ src/utils/network/network.ts
@@ -1,8 +1,7 @@
 import { ApiPromise } from '@polkadot/api';
 import { chain as lodashChain, curry, curryRight, has, isEqual, isNull, omit, once, pick, upperFirst } from 'lodash';
 import Web3 from 'web3';
-import { getCustomNetworkConfig, getUnit } from '../helper';
-import { SYSTEM_NETWORK_CONFIGURATIONS, NETWORK_SIMPLE, tronConfig } from '../../config';
+import { NETWORK_SIMPLE, SYSTEM_NETWORK_CONFIGURATIONS, tronConfig } from '../../config/network';
 import {
   Arrival,
   ChainConfig,
@@ -21,6 +20,8 @@ import {
   PolkadotConnection,
   Vertices,
 } from '../../model';
+import { getUnit } from '../helper/balance';
+import { getCustomNetworkConfig } from '../helper/storage';
 import { entrance } from './entrance';
 import { AIRDROP_GRAPH, NETWORK_GRAPH } from './graph';
