+++ src/utils/token/tokenInfo.ts
@@ -6,7 +6,7 @@ import { ApiPromise } from '@polkadot/api';
 import { memoize } from 'lodash';
 import Web3 from 'web3';
 import { AbiItem } from 'web3-utils';
-import { abi } from '../../config';
+import { abi } from '../../config/abi';
 import { entrance, waitUntilConnected } from '../network';
 
 export interface TokenCache {
